Moving cservice code around
